2009-08-26  Atsushi Enomoto  <atsushi@ximian.com>
	* PeerNode.cs : use ulong as same as Welcome message value.

	* PeerDuplexChannel.cs : in the Welcome response, return the NodeId
	  of the receiver, not the requestor. It resolved the issue that
	  .NET client didn't send further messages to mono server.


svn path=/trunk/mcs/; revision=140693

+++ mcs/class/System.ServiceModel/System.ServiceModel/PeerNode.cs
@@ -49,7 +49,7 @@ namespace System.ServiceModel
 
 		internal string MeshId { get; private set; }
 
-		internal int NodeId { get; set; }
+		internal ulong NodeId { get; set; }
 
 		internal bool IsOpen {
 			get { return RegisteredId != null; }
